// frontend/src/app/core/api/model/monthly-holding-entry.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MonthlyHoldingEntry { 
    name?: string;
    tickers?: Array<string>;
    marketValueEur?: number;
    costBasisEur?: number;
    unrealizedPlEur?: number;
    unrealizedPlPercentage?: number;
    portfolioPercentage?: number;
}