// frontend/src/app/core/api/model/monthly-invested-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MonthlyInvestedEntry } from './monthly-invested-entry';


export interface MonthlyInvestedResponse { 
    series?: Array<MonthlyInvestedEntry>;
}