// frontend/src/app/core/api/model/monthly-pl-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MonthlyPlEntry } from './monthly-pl-entry';


export interface MonthlyPlResponse { 
    series?: { [key: string]: Array<MonthlyPlEntry>; };
}