// frontend/src/app/core/api/model/portfolio-holding-monthly-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MonthlyHoldingEntry } from './monthly-holding-entry';


export interface PortfolioHoldingMonthlyResponse { 
    yearMonth?: string;
    holdings?: Array<MonthlyHoldingEntry>;
}